Organize file path of import
