+++ src/components/widgets/currencies.tsx
@@ -14,31 +14,31 @@ import "/node_modules/react-resizable/css/styles.css"
 import { DocumentData } from "@firebase/firestore-types"
 import { db } from "../../utils/firebase"
 import { doc, getDoc } from "firebase/firestore"
-import usa from "../assets/flags/usa.png"
-import canada from "../assets/flags/canada.png"
-import japan from "../assets/flags/japan.png"
-import euro from "../assets/flags/euro.png"
-import uk from "../assets/flags/uk.png"
-import krw from "../assets/flags/sKorea.png"
-import china from "../assets/flags/china.png"
-import taiwan from "../assets/flags/taiwan.png"
-import hkd from "../assets/flags/hkd.png"
-import macao from "../assets/flags/macao.png"
-import singapore from "../assets/flags/singapore.png"
-import thailand from "../assets/flags/thailand.png"
-import malaysia from "../assets/flags/malaysia.png"
-import vietnam from "../assets/flags/vietnam.png"
-import indonesia from "../assets/flags/indonesia.png"
-import philippine from "../assets/flags/philippine.png"
-import australia from "../assets/flags/australia.png"
-import newZealand from "../assets/flags/newZealand.png"
-import sweden from "../assets/flags/sweden.png"
-import switzerland from "../assets/flags/switzerland.png"
-import denmark from "../assets/flags/denmark.png"
-import mexico from "../assets/flags/mexico.png"
-import turkey from "../assets/flags/turkey.png"
-import india from "../assets/flags/india.png"
-import calculator from "../assets/calculator.png"
+import usa from "../../assets/flags/usa.png"
+import canada from "../../assets/flags/canada.png"
+import japan from "../../assets/flags/japan.png"
+import euro from "../../assets/flags/euro.png"
+import uk from "../../assets/flags/uk.png"
+import krw from "../../assets/flags/sKorea.png"
+import china from "../../assets/flags/china.png"
+import taiwan from "../../assets/flags/taiwan.png"
+import hkd from "../../assets/flags/hkd.png"
+import macao from "../../assets/flags/macao.png"
+import singapore from "../../assets/flags/singapore.png"
+import thailand from "../../assets/flags/thailand.png"
+import malaysia from "../../assets/flags/malaysia.png"
+import vietnam from "../../assets/flags/vietnam.png"
+import indonesia from "../../assets/flags/indonesia.png"
+import philippine from "../../assets/flags/philippine.png"
+import australia from "../../assets/flags/australia.png"
+import newZealand from "../../assets/flags/newZealand.png"
+import sweden from "../../assets/flags/sweden.png"
+import switzerland from "../../assets/flags/switzerland.png"
+import denmark from "../../assets/flags/denmark.png"
+import mexico from "../../assets/flags/mexico.png"
+import turkey from "../../assets/flags/turkey.png"
+import india from "../../assets/flags/india.png"
+import calculator from "../../assets/calculator.png"
 
 const Title = styled.div`
   padding: 0;
