+++ src/components/widgets/tools.tsx
@@ -3,15 +3,15 @@ import { useDrag } from "@use-gesture/react"
 import { useSpring, animated } from "@react-spring/web"
 import styled from "styled-components"
 import { DocumentData } from "@firebase/firestore-types"
-import robot from "../assets/chatbot.png"
-import currency from "../assets/whiteCurrencies.png"
-import questionMark from "../assets/question-mark.png"
-import weather from "../assets/whiteWeather.png"
+import robot from "../../assets/chatbot.png"
+import currency from "../../assets/whiteCurrencies.png"
+import questionMark from "../../assets/question-mark.png"
+import weather from "../../assets/whiteWeather.png"
 import CurrencyWidget, { getRatesData } from "./currencies"
 import WeatherWidget from "./weather"
 import { ToolContext } from "../../context/toolContext"
-import usa from "../assets/flags/usa.png"
-import taiwan from "../assets/flags/taiwan.png"
+import usa from "../../assets/flags/usa.png"
+import taiwan from "../../assets/flags/taiwan.png"
 
 const ToolsWrapper = styled.div`
   position: absolute;
